Update axis brushes for window resize
showResize(): heightChanged : redraw axis brushes.
updateRange(): set brush extents to changed yRange.
split maybeFlip(), maybeFlipExtent() out of draw-map.js to utils/draw/axis.js
frontend/app/
 202e995 221844 Jun 26 23:54  components/draw-map.js
 50cfa65   3010 Jun 26 22:01  utils/draw/axis.js
 da6760b   1216 Jun 26 23:46  utils/stacksLayout.js
